requirements.txt
# Core ML (Pinned for stability and matching DLLs)
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# Outras dependencias essenciais
numpy<2.1.0
g4f[all]
yt-dlp
ffmpeg-python
whisperx
mediapipe
google-generativeai
google-genai
insightface
gradio
opencv-python
psutil
fastapi
uvicorn
deep-translator
tqdm
sqlalchemy
